jw changes
tuned some of the language to make it clear, for example, that the circuits are specific to the Radish34 demo.  Added a paragraph to make it clear that new developers building things other than MSAs would have a few options...and that standards bodies have an opportunity to make libraries of specialized circuits.  Made a few other grammar edits.