Fix baseline_track rejected promise assertion in failure case
The current implementation requires a shield contract to exist at the
address provided to the baseline_trace(Address) RPC call. If there is
ever a reason to leverage CREATE2 or otherwise begin tracking a shield
contract address before the contract code has been deployed to the blockchain,
this will have to be revisited.